create table and database text file

+++ demo/src/main/java/com/example/demo/service/StudentService.java
@@ -49,7 +49,7 @@ public class StudentService {
 
 
     public String deleteStudent(int id) {
-        studentRepository.deleteById((long)id);
+        studentRepository.deleteById(id);
         return "deleted";
     }
 }
